Working serial version of the parameter extraction.
Weirdly, it does seem to do a fair bit of parallelism (dozens of threads
in places) even though I haven't written any yet. Not sure why. Presumably
It's at least partly due to some of the libraries I've pulled in to help
with this, but it's not obvious to me how or why that would be the result.