Implemented various attempts with transducers.
None of these made a huge improvement (about 83% of the "serial"
time was the best I got). And it's a lot of work to try to get these
to work (with the exception of the simple `pmap` version). I really
think that the `core.async` approach is likely to just be a Much
Better Idea (but I don't know what until it's done).